Add better method for determining direction of corotating frame
First, use new method in `quaternion` for integrating angular velocity
for tabulated data.  This avoids the need for `squad` by using dense
output to get more accurate values at each input time.

Second, "the" corotating frame is only unique up to an overall rotation.
This commit allows that rotation to be set by choosing a span of times
over which the dominant eigenvector of the corotating waveform's <LL>
matrix will be as close as possible to the z axis on average.