Make fixture-generating funcs work again after pytest deprectation (#27)
See
https://docs.pytest.org/en/latest/deprecations.html#calling-fixtures-directly
for details on the deprecation. The fixture-generating functions:
  constant_wavform
  linear_waveform
  random_waveform
  delta_waveform
were designed to be general and take optional arguments. Only their
default version should be considered as fixtures, because otherwise
they need to be called directly.  But, pytest has deprecated direct
calling of fixtures.  Therefore we make 0-argument fixtures (with
explicit names given to the decorator) which return the default values
from the above fixture-generating functions.  Then the underlying
functions can still be called with optional arguments.